+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.3](https://github.com/alloy-rs/nybbles/releases/tag/v0.4.3) - 2025-08-15
+
+### Bug Fixes
+
+- `pack_to_unchecked` on big-endian targets ([#35](https://github.com/alloy-rs/nybbles/issues/35))
+
+## [0.4.2](https://github.com/alloy-rs/nybbles/releases/tag/v0.4.2) - 2025-08-12
+
+### Features
+
+- Add iterator for Nibbles ([#34](https://github.com/alloy-rs/nybbles/issues/34))
+
+### Miscellaneous Tasks
+
+- Release 0.4.2
+
 ## [0.4.1](https://github.com/alloy-rs/nybbles/releases/tag/v0.4.1) - 2025-07-07
 
 ### Bug Fixes
@@ -14,6 +30,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Features
 
 - Add back support for big-endian targets ([#32](https://github.com/alloy-rs/nybbles/issues/32))
+
+### Miscellaneous Tasks
+
+- Release 0.4.1
 
 ## [0.4.0](https://github.com/alloy-rs/nybbles/releases/tag/v0.4.0) - 2025-06-19